Demo Initialization
This change encapsulates a set of changes that need to be made to the
demo application before the start of giving the demo.  This, along
with the speaker notes, describe the changes that need to be made to
the application as part of the flow of the talk.  The changes in made
in this commit to do not prevent the code from compiling, but rather
cause test failures.  Once the TODOs are completed, the tests should
pass and the application will be back to fully-functioning.